+++ extend/waibuDb/schema/secure-guard.js
@@ -1,0 +1,15 @@
+async function secureGuard () {
+  return {
+    common: {
+      widget: {
+        teamIds: {
+          attr: {
+            refUrl: 'waibuAdmin:/{prefix}/team/details?id={id}'
+          }
+        }
+      }
+    }
+  }
+}
+
+export default secureGuard