+++ extend/waibuMpa/extend/waibuAdmin/route/anonymous-guard/@action.js
@@ -1,11 +1,10 @@
 const action = {
   method: ['GET', 'POST'],
-  title: 'xAttribGuard',
-  xSite: true,
+  title: 'anonymousGuard',
   handler: async function (req, reply) {
     const { importModule } = this.app.bajo
     const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')
-    return await crudSkel.call(this, 'SumbaXAttribGuard', req, reply)
+    return await crudSkel.call(this, 'SumbaAnonymousGuard', req, reply)
   }
 }