+++ extend/waibuMpa/extend/waibuAdmin/route/secure-guard/@action.js
@@ -1,11 +1,10 @@
 const action = {
   method: ['GET', 'POST'],
-  title: 'xRouteGuard',
-  xSite: true,
+  title: 'secureGuard',
   handler: async function (req, reply) {
     const { importModule } = this.app.bajo
     const crudSkel = await importModule('waibuAdmin:/lib/crud-skel.js')
-    return await crudSkel.call(this, 'SumbaXRouteGuard', req, reply)
+    return await crudSkel.call(this, 'SumbaSecureGuard', req, reply)
   }
 }