+++ extend/waibuMpa/route/access-token.js
@@ -1,7 +1,9 @@
 const apiToken = {
   method: 'POST',
   handler: async function (req, reply) {
-    if (!req.user) return ''
+    const { get } = this.app.lib._
+    const uid = get(req, 'user.id')
+    if (!uid) return ''
     const rec = await this.app.dobo.getModel('SumbaUser').getRecord(req.user.id, { forceNoHidden: true, noCache: true })
     return (await this.createJwtFromUserRecord(rec)).token
   }